requirements.txt
absl-py==0.9.0
aiofiles==0.5.0
aiohttp==3.6.2
APScheduler==3.6.3
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
blis==0.4.1
boto3==1.14.47
botocore==1.17.47
cachetools==4.0.0
catalogue==1.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
fbmessenger==6.0.0
future==0.18.2
gast==0.2.2
gevent==1.5.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
greenlet==0.4.16
grpcio==1.31.0
h11==0.8.1
h2==3.2.0
h5py==2.10.0
hpack==3.0.0
hstspreload==2020.8.18
httplib2==0.17.0
httptools==0.1.1
httpx==0.9.3
humanfriendly==8.2
hyperframe==5.2.0
idna==2.9
importlib-metadata==1.7.0
jmespath==0.10.0
joblib==0.16.0
jsonpickle==1.4.1
jsonschema==3.2.0
kafka-python==1.4.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.2
mattermostwrapper==2.2
multidict==4.7.5
murmurhash==1.0.2
networkx==2.4
nltk==3.4.5
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.2
pika==1.1.0
plac==1.1.3
preshed==3.0.2
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.8.0
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.12.1
python-socketio==4.5.1
python-telegram-bot==12.4.2
pytz==2019.3
PyYAML==5.3.1
questionary==1.5.2
rake-nltk==1.0.4
rasa==1.10.10
rasa-sdk==1.10.2
redis==3.5.3
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.3.1
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
sanic==19.12.2
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.3
scikit-learn==0.22.2.post1
scipy==1.5.2
six==1.14.0
sklearn-crfsuite==0.3.6
slack-client==0.3.0
slackclient==2.5.0
sniffio==1.1.0
spacy==2.3.2
SQLAlchemy==1.3.15
srsly==1.0.2
tabulate==0.8.7
tensorboard==2.1.1
tensorflow==2.1.1
tensorflow-addons==0.7.1
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
tensorflow-probability==0.9.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.4.1
tornado==6.0.4
tqdm==4.45.0
twilio==6.26.3
tzlocal==2.1
ujson==2.0.3
uritemplate==3.0.1
urllib3==1.25.8
uvloop==0.14.0
wasabi==0.7.1
wcwidth==0.2.5
webexteamssdk==1.3
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.4.2
zipp==3.1.0
